requirements.txt
telethon>=1.24.0
pyyaml>=6.0 